+++ instrumentation/opentelemetry-instrumentation-django/tests/test_middleware.py
@@ -22,6 +22,7 @@ from django.http import HttpRequest, HttpResponse
 from django.test.client import Client
 from django.test.utils import setup_test_environment, teardown_test_environment
 
+from opentelemetry import trace
 from opentelemetry.instrumentation.django import (
     DjangoInstrumentor,
     _DjangoMiddleware,
@@ -409,6 +410,7 @@ class TestMiddlewareWithTracerProvider(TestBase, WsgiTestBase):
         result = self.create_tracer_provider(resource=resource)
         tracer_provider, exporter = result
         self.exporter = exporter
+        self.tracer_provider = tracer_provider
         _django_instrumentor.instrument(tracer_provider=tracer_provider)
 
     def tearDown(self):
@@ -432,3 +434,20 @@ class TestMiddlewareWithTracerProvider(TestBase, WsgiTestBase):
         self.assertEqual(
             span.resource.attributes["resource-key"], "resource-value"
         )
+
+    def test_django_with_wsgi_instrumented(self):
+        tracer = self.tracer_provider.get_tracer(__name__)
+        with tracer.start_as_current_span(
+            "test", kind=SpanKind.SERVER
+        ) as parent_span:
+            Client().get("/span_name/1234/")
+            span_list = self.exporter.get_finished_spans()
+            print(span_list)
+            self.assertEqual(
+                span_list[0].attributes[SpanAttributes.HTTP_STATUS_CODE], 200
+            )
+            self.assertEqual(trace.SpanKind.INTERNAL, span_list[0].kind)
+            self.assertEqual(
+                parent_span.get_span_context().span_id,
+                span_list[0].parent.span_id,
+            )